requirements.txt
psycopg2-binary==2.9.9
requests==2.32.3
Flask-SQLAlchemy==3.1.1
pydantic-settings==2.4.0